배포 1.0.1 : 미디어쿼리 최적화

+++ product/views.py
@@ -1,5 +1,3 @@
-from itertools import chain
-
 from django.db.models import Sum, F
 from django.shortcuts import render
 from django.views.decorators.http import require_POST
@@ -13,155 +11,52 @@ import member.urls
 import datetime
 from django.core import serializers
 import json
-import itertools
+
 
 def index(request):
-    # 각 카테고리에서 판매량 1순위인 상품 리스트에 추가
-    # MEN, WOMEN 구분 함.
-    product_list = []
-    category = Category.objects.all()
-    count = range(1, len(category)+1)
-    pro3 = Product.objects.filter(gender__isnull=True)
-    for i in count:
-        # pro: men의 category_id가 i인 상품 중 판매량 1순위인 상품 1개
-        # pro2: women의 category_id가 i인 상품 중 판매량 1순위인 상품 1개
-        # pro3: men과 women의 각 카테고리에서 판매량 1순위인 상품 최종 리스트
-        pro = Product.objects.filter(gender='MEN', category_id=i).order_by('-sales')[:1]
-        pro2 = Product.objects.filter(gender='WOMEN', category_id=i).order_by('-sales')[:1]
-        pro3 = itertools.chain(pro, pro2, pro3)
-    product_list = pro3
     return render(
         request,
         'product/index.html/',
-        {'product_list': product_list},
+        {},
     )
-
-
 def about(request):
     return render(
         request,
         'product/about.html/',
         {},
     )
-
-
 def error(request):
     return render(
         request,
         'product/error.html/',
         {},
     )
-
-
-def FilterDetail(request):
+def SizeDetail(request):
     template_name = 'product/product.html'
     if request.method == "POST":
         size_list = request.POST.getlist('size')
-        align = request.POST.get('orderType')
         url = request.POST.get('url')
         url = url.split('/')[3:5]
-
         size_id = []
         for i in size_list:
-            product = Inventory.objects.filter(size=i, amount__gt=0).values_list('product_id', flat=True)
+            product = Inventory.objects.filter(size=i, soldout=False).values_list('product_id', flat=True)
             for pro in product:
                 size_id.append(pro)
-
-        # 사이즈 필터 선택 유무 flag
-        if not size_id:
-            size_flag = 0
-        else:
-            size_flag = 1
-
-        # 정렬 방식 선택 유무 flag
-        if align is None:
-            align_flag = 0
-        else:
-            align_flag = 1
-
-        print(size_flag)
-        print(align_flag)
-        print(size_id)
-
         if url[0] == '1':
-            # MEN 카테고리
             gender = 'Men'
             if url[1] == '0':
-                # 신발 전체
-                if size_flag == 1:
-                    # 사이즈 필터 ON
-                    if align_flag == 1:
-                        # 정렬 방식 선택(align: 정렬 방식)
-                        queryset = Product.objects.filter(gender='MEN', pk__in=size_id)
-                        product_list = serializers.serialize("json", FilterList(align, queryset))
-                    else:
-                        product_list = serializers.serialize("json",
-                                                             Product.objects.filter(gender='MEN', pk__in=size_id))
-                else:
-                    # 사이즈 필터 OFF
-                    if align_flag == 1:
-                        # 정렬 방식 선택(align: 정렬 방식)
-                        queryset = Product.objects.filter(gender='MEN')
-                        product_list = serializers.serialize("json", FilterList(align, queryset))
-                    else:
-                        product_list = serializers.serialize("json", Product.objects.filter(gender='MEN'))
+                product_list = serializers.serialize("json", Product.objects.filter(gender='MEN', pk__in=size_id))
                 category = '신발'
             else:
-                # 카테고리 별 신발
-                if size_flag == 1:
-                    if align_flag == 1:
-                        queryset = Product.objects.filter(gender='MEN', category_id=url[1], pk__in=size_id)
-                        product_list = serializers.serialize("json", FilterList(align, queryset))
-                    else:
-                        product_list = serializers.serialize("json",
-                                                             Product.objects.filter(gender='MEN', category_id=url[1],
-                                                                                    pk__in=size_id))
-                else:
-                    if align_flag == 1:
-                        queryset = Product.objects.filter(gender='MEN', category_id=url[1])
-                        product_list = serializers.serialize("json", FilterList(align, queryset))
-                    else:
-                        product_list = serializers.serialize("json", Product.objects.filter(gender='MEN', category_id=url[1]))
+                product_list = serializers.serialize("json", Product.objects.filter(gender='MEN', category_id=url[1], pk__in=size_id))
                 category = serializers.serialize("json", Category.objects.filter(pk=url[1]), fields=('name'))
         else:
-            # WOMEN 카테고리
             gender = 'Women'
             if url[1] == '0':
-                # 신발 전체
-                if size_flag == 1:
-                    # 사이즈 필터 ON
-                    if align_flag == 1:
-                        # 정렬 방식 선택(align: 정렬 방식)
-                        queryset = Product.objects.filter(gender='WOMEN', pk__in=size_id)
-                        product_list = serializers.serialize("json", FilterList(align, queryset))
-                    else:
-                        product_list = serializers.serialize("json",
-                                                             Product.objects.filter(gender='WOMEN', pk__in=size_id))
-                else:
-                    # 사이즈 필터 OFF
-                    if align_flag == 1:
-                        queryset = Product.objects.filter(gender='WOMEN')
-                        product_list = serializers.serialize("json", FilterList(align, queryset))
-                    else:
-                        product_list = serializers.serialize("json", Product.objects.filter(gender='WOMEN'))
+                product_list = serializers.serialize("json", Product.objects.filter(gender='WOMEN', pk__in=size_id))
                 category = '신발'
             else:
-                # 카테고리 별 신발
-                if size_flag == 1:
-                    if align_flag == 1:
-                        queryset = Product.objects.filter(gender='WOMEN', category_id=url[1], pk__in=size_id)
-                        product_list = serializers.serialize("json", FilterList(align, queryset))
-                    else:
-                        product_list = serializers.serialize("json",
-                                                             Product.objects.filter(gender='WOMEN', category_id=url[1],
-                                                                                    pk__in=size_id))
-                else:
-                    if align_flag == 1:
-                        queryset = Product.objects.filter(gender='WOMEN', category_id=url[1])
-                        product_list = serializers.serialize("json", FilterList(align, queryset))
-                    else:
-                        product_list = serializers.serialize("json",
-                                                             Product.objects.filter(gender='WOMEN', category_id=url[1]))
+                product_list = serializers.serialize("json", Product.objects.filter(gender='WOMEN', category_id=url[1], pk__in=size_id))
                 category = serializers.serialize("json", Category.objects.filter(pk=url[1]), fields=('name'))
 
         ret = {'product_list': product_list,
@@ -169,19 +64,6 @@ def FilterDetail(request):
                'category': category}
 
         return HttpResponse(json.dumps(ret), content_type='application/json')
-
-
-def FilterList(align, queryset):
-    # 선택 된 정렬 방식에 따라 상품 align
-
-    if align == 'new-order':
-        product_list = queryset.order_by('-release_date')
-    elif align == 'low-order':
-        product_list = queryset.order_by('price')
-    else:
-        product_list = queryset.order_by('-price')
-
-    return product_list
 
 
 class CategoryDetail(ListView):
@@ -211,7 +93,6 @@ class CategoryDetail(ListView):
 
         return context
 
-
 class NewProductList(ListView):
     # 모든 신발 카테고리 해당. today 기준 출시일이 30일 전 이내인 상품만
     model = Product
@@ -230,13 +111,11 @@ class NewProductList(ListView):
                 gender='MEN',
                 release_date__range=[one_month_ago, now_date]
             ).reverse()
-            context['gender'] = 'Men'
         else:
             context['product_list'] = Product.objects.filter(
                 gender='WOMEN',
                 release_date__range=[one_month_ago, now_date]
             ).reverse()
-            context['gender'] = 'Women'
 
         return context
 
@@ -257,7 +136,6 @@ class BestProductList(ListView):
             context['bask_product_list'] = Product.objects.filter(gender='MEN', category_id=3).order_by('-sales')[:5]
             context['soc_product_list'] = Product.objects.filter(gender='MEN', category_id=4).order_by('-sales')[:5]
             context['flip_product_list'] = Product.objects.filter(gender='MEN', category_id=5).order_by('-sales')[:5]
-            context['gender'] = 1
         else:
             # WOMEN의 BEST 품목
             context['life_product_list'] = Product.objects.filter(gender='WOMEN', category_id=1).order_by('-sales')[:5]
@@ -265,71 +143,43 @@ class BestProductList(ListView):
             context['bask_product_list'] = Product.objects.filter(gender='WOMEN', category_id=3).order_by('-sales')[:5]
             context['soc_product_list'] = Product.objects.filter(gender='WOMEN', category_id=4).order_by('-sales')[:5]
             context['flip_product_list'] = Product.objects.filter(gender='WOMEN', category_id=5).order_by('-sales')[:5]
-            context['gender'] = 2
         return context
 
 
 class SaleProductList(ListView):
     model = Product
-    template_name = 'product/product-sales.html'
+    template_name = 'product/product.html'
 
     def get_context_data(self, **kwargs):
         context = super(SaleProductList, self).get_context_data(**kwargs)
-        id = self.kwargs['id']
-        gender = self.kwargs['gender']
+        pk = self.kwargs['pk']
 
         date_format = "%Y-%m-%d"
         three_month_ago = (datetime.datetime.now() - datetime.timedelta(days=90)).strftime(date_format)
         now_date = datetime.datetime.now().strftime(date_format)
 
-        url = []
-        category = []
-
-        # 카테고리 별 url & 카테고리 name 리스트
-        for i in range(1, 6):
-            category_id = Category.objects.filter(pk=i).values_list('name', flat=True)
-            for k in category_id:
-                category.append(k)
-            for j in range(1, 3):
-                url.append("{% url 'products:sale' " + str(j) + ' ' + str(i) + " %}")
-        context['url'] = url
-        context['category'] = category
-
-        # 카테고리 별 sale 상품 리스트
-        if gender == 1:
-            context['gender'] = 'Men'
-            if id == 0:
-                context['product_list'] = Product.objects.filter(
-                    gender='MEN',
-                    release_date__lte=three_month_ago,
-                    soldout=False
-                )
-            else:
-                context['category_id'] = id
-                context['product_list'] = Product.objects.filter(
-                    gender='MEN',
-                    release_date__lt=three_month_ago,
-                    soldout=False,
-                    category_id=id
-                ).reverse()
+        if pk == 1:
+            context['product_list'] = Product.objects.filter(
+                gender='MEN',
+                release_date__lt=three_month_ago,
+                soldout=False
+            ).reverse()
         else:
-            context['gender'] = 'Women'
-            if id == 0:
-                context['product_list'] = Product.objects.filter(
-                    gender='WOMEN',
-                    release_date__lt=three_month_ago,
-                    soldout=False
-                ).reverse()
-            else:
-                context['category_id'] = id
-                context['product_list'] = Product.objects.filter(
-                    gender='WOMEN',
-                    release_date__lt=three_month_ago,
-                    soldout=False,
-                    category_id=id
-                ).reverse()
+            context['product_list'] = Product.objects.filter(
+                gender='WOMEN',
+                release_date__lt=three_month_ago,
+                soldout=False
+            ).reverse()
 
         return context
+
+
+# def cart(request):
+#     return render(request, 'product/cart.html', {})
+
+
+def detail(request):
+    return render(request, 'product/detail.html', {})
 
 class ProductDetail(DetailView):
     model = Product
@@ -391,10 +241,12 @@ def cart_delete_one(request):      # 장바구니 특정상품 삭제
     cart_id = request.POST['cart-id']
     data = Cart.objects.get(id=cart_id)     # id가 cart_id와 일치하는 쿼리셋 반환
 
-    # 삭제 성공
-    data.delete()
-    request.session['cart_count'] -= 1
-    return HttpResponse(json.dumps({'result': 'success'}), content_type="application/json")
+    if data.user_id.id == user_id:       # 삭제 성공
+        data.delete()
+        request.session['cart_count'] -= 1
+        return HttpResponse(json.dumps({'result': 'success'}), content_type="application/json")
+    else:       # Cart의 user_id와 로그인한 유저가 다른 경우
+        return HttpResponse(json.dumps({'result': 'no action'}), content_type="application/json")
 
 
 def cart_delete_all(request):      # 장바구니 전체상품 삭제
@@ -403,6 +255,10 @@ def cart_delete_all(request):      # 장바구니 전체상품 삭제
         return HttpResponse(json.dumps({'result': 'no user'}), content_type="application/json")
 
     user_id = request.user.pk
+
+    # Cart의 user_id와 로그인한 유저가 다른 경우
+    if int(request.POST['user-id']) != user_id:
+        return HttpResponse(json.dumps({'result': 'no action'}), content_type="application/json")
 
     data = Cart.objects.filter(user_id=user_id)     # user_id가 일치하는 쿼리셋 반환
     data.delete()
@@ -428,22 +284,11 @@ def get_option(request, **kwargs):
     for i in product_image:
         image.append(i.image.url)
     for i in inventory_set:
-        inventory.append({'id': i.id, 'size': i.size, 'amount': i.amount})
+        inventory.append({'size': i.size, 'amount': i.amount})
 
     option = {'name': name, 'category': category, 'price': price, 'image': image, 'inventory': inventory}
 
     return HttpResponse(json.dumps({'result': 'success', 'option': option}), content_type="application/json")
-
-
-def change_option(request):
-    cart_id = request.POST['cart-id']
-    inventory_id = request.POST['inventory-id']
-
-    cart = Cart.objects.get(pk=cart_id)
-    cart.inventory_id = Inventory.objects.get(pk=inventory_id)
-    cart.save()
-
-    return HttpResponse(json.dumps({'result': 'success'}), content_type="application/json")
 
 
 class CartList(LoginRequiredMixin, ListView):
